+++ src/shape/nodes/diamond.ts
@@ -56,8 +56,8 @@ Shape.registerNode(
 
       const keyShape = group.addShape('path', {
         attrs: style,
-        className: 'diamond-keyShape',
-        name: 'diamond-keyShape',
+        className: `${this.type}-keyShape`,
+        name: `${this.type}-keyShape`,
         draggable: true,
       });
 
@@ -69,8 +69,8 @@ Shape.registerNode(
             y: -h! / 2,
             ...icon,
           },
-          className: 'diamond-icon',
-          name: 'diamond-icon',
+          className: `${this.type}-icon`,
+          name: `${this.type}-icon`,
           draggable: true
         });
       }